Share thread resolution and execute body across channel schedule adapters
Lark and weixin schedule adapters were ~95% verbatim. Hoist supports,
execute, resolveThread, threadExists, preferredAgentFor into a new
BaseChannelScheduleAdapter; each platform now contributes only its
policy factory, route types, and delivery target.
